// src/wkof_types.d.ts
type SubjectType = 'vocabulary' | 'kanji' | 'radical'
type SubjectTypeString =
  | `${SubjectType}`
  | `${SubjectType},${SubjectType}`
  | `${SubjectType},${SubjectType},${SubjectType}`
type SubjectTypeShort = 'voc' | 'kan' | 'rad'
type SubjectTypeShortString =
  | `${SubjectTypeShort}`
  | `${SubjectTypeShort},${SubjectTypeShort}`
  | `${SubjectTypeShort},${SubjectTypeShort},${SubjectTypeShort}`
type IsoDateString = `${number}-${number}-${number}T${number}:${number}:${number}.${number}Z`

declare namespace Core {
  type FileCache = {
    dir: {
      [key: string]: {
        added: IsoDateString
        last_loaded: IsoDateString
      }
    }
    save: (name: string, content: string | { [key: string]: any }) => Promise<string>
    load: (name: string) => Promise<string | { [key: string]: any }>
    delete: (name: string | RegExp) => Promise<undefined>
    clear: () => Promise<undefined>
  }

  export type Module = {
    version: {
      value: string
      compare_to: (needed_version: string) => 'older' | 'same' | 'newer'
    }
    include: (modules: string) => Promise<{ loaded: string[]; failed: string[] }>
    ready: (modules: string) => Promise<'ready'>
    get_state: (state_var: string) => any
    set_state: (state_var: string, value: any) => void
    wait_state: (
      state_var: string,
      value: any,
      callback: (new_value: any, prev_value: any) => void,
      persistent: boolean,
    ) => void
    on: (event: string, callback: () => void) => void
    trigger: (event: string) => undefined
    support_files: { [key: string]: string }
    load_file: (url: string, use_cache?: boolean) => Promise<any>
    load_script: (url: string, use_cache?: boolean) => Promise<string | undefined>
    load_css: (url: string, use_cache?: boolean) => Promise<string | undefined>
    file_cache: FileCache
  }
}

declare namespace Apiv2 {
  type ApiEndpoint =
    | 'assignments'
    | 'level_progressions'
    | 'resets'
    | 'reviews'
    | 'review_statistics'
    | 'spaced_repetition_systems'
    | 'study_materials'
    | 'subjects'
    | 'summary'
    | 'user'
    | 'voice_actors'

  type Defaults = {
    /**
     * @remark Comma separated list of integers
     */
    ids?: string
    updated_after?: IsoDateString
  }

  type LevelProgressions = {} & Defaults
  type Resets = {} & Defaults
  type SpacedRepetitionSystems = {} & Defaults
  type VoiceActors = {} & Defaults

  type Assignments = {
    available_after?: IsoDateString
    available_before?: IsoDateString
    burned?: boolean
    hidden?: boolean
    /**
     * @remark Value not needed
     */
    immediately_available_for_lessons?: any
    /**
     * @remark Value not needed
     */
    immediately_available_for_review?: any
    /**
     * @remark Value not needed
     */
    in_review?: any
    /**
     * @remark Comma separated list of integers
     */
    levels?: string
    /**
     * @remark Comma separated list of integers
     */
    srs_stages?: string
    started?: boolean
    /**
     * @remark Comma separated list of integers
     */
    subject_ids?: string
    subject_types?: SubjectTypeString
    unlocked?: boolean
  } & Defaults

  type Reviews = {
    assignment_id?: number
    created_at?: IsoDateString
    ending_srs_stage?: number
    incorrect_meaning_answers?: number
    incorrect_reading_answers?: number
    spaced_repetition_system_id?: number
    starting_srs_stage?: number
    subject_id?: number
  }

  type ReviewStatistics = {
    percentages_greater_than?: number
    percentages_less_than?: number
    /**
     * @remark Comma separated list of integers
     */
    subject_ids?: string
  } & Defaults

  type StudyMaterials = {
    hidden?: boolean
    /**
     * @remark Comma separated list of integers
     */
    subject_ids?: string
    subject_types?: SubjectTypeString
  } & Defaults

  type Subjects = {
    types?: SubjectTypeString
    /**
     * @remark Comma separated list of slugs
     */
    slugs?: string
    /**
     * @remark Comma separated list of levels
     */
    levels?: string
    hidden?: boolean
  } & Defaults

  type Filter =
    | Assignments
    | LevelProgressions
    | Resets
    | Reviews
    | ReviewStatistics
    | SpacedRepetitionSystems
    | StudyMaterials
    | Subjects
    | VoiceActors

  type FetchEndpointOptions = {
    progress_callback?: (endpoint_name: ApiEndpoint, first_new: number, so_far: number, total: number) => void
    last_update?: Date | IsoDateString
    filters?: Filter
  }

  type GetEndpointOptions = {
    progress_callback?: (endpoint_name: ApiEndpoint, first_new: number, so_far: number, total: number) => void
    force_update?: boolean
  }
  type User = {
    apikey: string
    current_vacation_started_at: null | IsoDateString
    id: string
    level: number
    preferences: {
      default_voice_actor: number
      extra_study_autoplay_audio: boolean
      lessons_autoplay_audio: boolean
      lessons_batch_size: number
      lessons_presentation_order: string
      reviews_autoplay_audio: boolean
      reviews_display_srs_indicator: boolean
      wanikani_compatibility_mode: boolean
    }
    profile_url: string
    started_at: IsoDateString
    subscription: {
      active: boolean
      max_level_granted: number
      period_ends_at: null | IsoDateString
      type: string
    }
    username: string
  }

  export type Module = {
    Apiv2: {
      user: string
      key: string
      fetch_endpoint: (endpoint_name: ApiEndpoint, options?: FetchEndpointOptions) => Promise<any>
      get_endpoint: (endpoint_name: ApiEndpoint, options?: GetEndpointOptions) => Promise<any>
      clear_cache: (include_non_user?: boolean) => Promise<undefined>
      apiv2_is_valid_apikey_format: (apikey: string) => boolean
      spoof: (key: string) => void
    }
    user: User
  }
}

declare namespace ItemData {
// ...
  namespace GetItems {
    type Endpoint = 'subjects' | 'assignments' | 'review_statistics' | 'study_materials'
    type EndpointString =
      | `${Endpoint}`
      | `${Endpoint},${Endpoint}`
      | `${Endpoint},${Endpoint},${Endpoint}`
      | `${Endpoint},${Endpoint},${Endpoint},${Endpoint}`
// ...
    type SourceConfig = {
      options?: Options
      filters?: Filters
    }

    type Config = EndpointString | { [key: string]: SourceConfig }
  }
// ...
}

declare namespace Menu {
  export type Module = {
    Menu: {
      insert_script_link: (config: {
        name: string
        submenu?: string
        title: string
        class?: string
        on_click: (event: any) => void
      }) => void
    }
  }
}

declare namespace Settings {
  type OnChange = (name: string, value: any, config: Component) => void

  type Validate = (value: any, config: Component) => boolean | string | { valid: boolean; msg: string }

  type ComponentDefaults = {
    label?: string
    hover_tip?: string
    full_width?: boolean
    validate?: Validate
    on_change?: OnChange
    path?: string
    refresh_on_change?: boolean
  }

  type Section = {
    type: 'section'
    label?: string
  }

  type Divider = {
    type: 'divider'
  }

  type Group = {
    type: 'group'
    label?: string
    content: { [key: string]: Component }
  }

  type List = {
    type: 'list'
    size?: number
    content: { [key: string]: string }
  } & ({ multi?: false; default?: string } | { multi?: true; default: { [key: string]: boolean } }) &
    ComponentDefaults

  type Dropdown = {
    type: 'dropdown'
    default?: string
    content: { [key: string]: string }
  } & ComponentDefaults

  type Checkbox = {
    type: 'checkbox'
    default?: boolean
  } & ComponentDefaults

  type Input = {
    type: 'input'
    subtype?: string
    default: string
  } & ComponentDefaults

  type Number = {
    type: 'number'
    placeholder?: string
    default?: number
    min?: number
    max?: number
  } & ComponentDefaults

  type Text = {
    type: 'text'
    placeholder?: string
    match?: RegExp
  } & ComponentDefaults

  type Color = {
    type: 'color'
    default?: string
  } & ComponentDefaults

  type Button = {
    type: 'button'
    text?: string
    on_click: (name: string, config: Component, on_change: () => void) => void
  } & ComponentDefaults

  type Html = {
    type: 'html'
    label?: string
    html: string
  }

  type Page = {
    type: 'page'
    label: string
    hover_tip?: string
    content: { [key: string]: Component }
  }

  type Tabset = {
    type: 'tabset'
    content: { [key: string]: Page }
  }

  type Component =
    | Tabset
    | Page
    | Section
    | Divider
    | Group
    | List
    | Dropdown
    | Checkbox
    | Input
    | Number
    | Text
    | Color
    | Button
    | Html

  type Dialog = {
    open: () => Promise<undefined>
    load: () => Promise<{ [key: string]: any }>
    save: () => Promise<undefined>
    refresh: () => void
    background: {
      open: () => void
      close: () => void
    }
    close: (save_pending_changes?: boolean) => void
    cancel: () => void
  }

  type Config = {
    script_id: string
    title: string
    autosave?: boolean
    background?: boolean
    no_bkgd?: boolean
    /**
     * @param dialog jQuery DOM element of dialog box
     */
    // @ts-ignore
    pre_open?: (dialog: JQuery) => void
    on_save?: (settings: { [key: string]: any }) => void
    on_cancel?: (settings: { [key: string]: any }) => void
    on_close?: (settings: { [key: string]: any }) => void
    on_change?: OnChange
    on_refresh?: (settings: { [key: string]: any }) => void
    content: { [key: string]: Component }
  }

  export type Module = {
    Settings: {
      new(config: Config): Dialog
      save: (script_id: string) => Promise<undefined>
      load: (script_id: string, defaults?: { [key: string]: any }) => Promise<{ [key: string]: any }>
    }
    settings: {
      [key: string]: {
        [key: string]: any
      }
    }
  }
}

declare namespace Progress {
  export type Module = {
    Progress: {
      update: (progress: { name: string; label: string; value: number; max: number }) => void
    }
  }
}

/**
 * @remark To include modules use WKOF & Module
 */
export type WKOF = Core.Module
export type Apiv2 = Apiv2.Module
export type ItemData = ItemData.Module
export type Menu = Menu.Module
export type Settings = Settings.Module